feat: add configuration for filesystem disks and storage symlinks

+++ config/filesystems.php
@@ -65,7 +65,7 @@ return [
             'host' => env('SFTP_HOST'),
             'username' => env('SFTP_USERNAME'),
             'password' => env('SFTP_PASSWORD'),
-            'port' => env('SFTP_PORT', 22),
+            'port' => (int) env('SFTP_PORT', 22),
             'root' => env('SFTP_ROOT', '/'),
             'timeout' => 30,
         ],
